+++ lib/plausible/stats/metrics.ex
@@ -26,6 +26,25 @@ defmodule Plausible.Stats.Metrics do
 
   def metric?(value), do: Enum.member?(@all_metrics, value)
 
+  on_ee do
+    def default_value(metric, query, dimensions)
+        when metric in [:average_revenue, :total_revenue],
+        do: Plausible.Stats.Goal.Revenue.format_revenue_metric(nil, query, dimensions)
+  end
+
+  def default_value(:visit_duration, _query, _dimensions), do: nil
+  def default_value(:scroll_depth, _query, _dimensions), do: nil
+
+  @float_metrics [
+    :views_per_visit,
+    :bounce_rate,
+    :percentage,
+    :conversion_rate,
+    :group_conversion_rate
+  ]
+  def default_value(metric, _query, _dimensions) when metric in @float_metrics, do: 0.0
+  def default_value(_metric, _query, _dimensions), do: 0
+
   def from_string!(str) do
     Map.fetch!(@metric_mappings, str)
   end